src/qronos/requirements-frozen-py2.txt
asn1crypto==0.24.0
cryptography==2.1.4
cvxopt==1.2.3
cvxpy==1.0.25
dill==0.3.1.1
ecos==2.0.7.post1
enum34==1.1.6
future==0.18.2
gmpy2==2.0.8
idna==2.6
ipaddress==1.0.17
keyring==10.6.0
keyrings.alt==3.0
mpmath==1.0.0
multiprocess==0.70.9
numpy==1.16.5
osqp==0.6.1
pycrypto==2.6.1
pygobject==3.26.1
pyxdg==0.25
repoze.lru==0.7
scipy==1.2.2
scs==2.1.1.post2
SecretStorage==2.3.1
six==1.11.0
